requirements.txt
pymysql==1.1.0
sqlalchemy==1.4
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==1.0.0
pyjwt==2.8.0